Moved routes that do not need access to rpc.rpc self into their own common file.
This is to reduce file size and separate api server routes with
privilege to access rpc.rpc defs and those that do not need access,
so should not.
